+++ CHANGES.rst
@@ -2,6 +2,7 @@ v0.0.2
 ######
 - Require |aiohttp[speedups]|_
 - Implement :py:func:`discover_devices <pupil_labs.realtime_api.discovery.discover_devices>`
+- Implement :py:class:`Control <pupil_labs.realtime_api.control.Control>`
 
 .. |aiohttp[speedups]| replace:: ``aiohttp[speedups]``
 .. _aiohttp[speedups]: https://docs.aiohttp.org/en/stable/